Share doc comment finding code
The code was duplicated between scripting and workspaces because it is
partially non-portable.  Share the files by linking (and, unfortunately,
ifdef'ing) instead.
